requirements.txt
torch>=2.0.0
numpy>=1.20.0
python-chess>=1.9.0
h5py>=3.7.0
PyYAML>=6.0 